test(sdk): move appliesToState interop tests into assertion_test.go
Fold the interop regression tests into the existing assertion test file
rather than carrying a separate assertion_interop_test.go. No change to
test content or outcomes.
